requirements.txt
snowflake-connector-python[pandas]
python-dotenv
openai
tqdm
scikit-learn